solo_project/youtube_viral/requirements.txt
google-api-python-client
pandas
numpy
supabase
python-dotenv
httpx
apscheduler
scikit-learn
